subs reducer: Remove needless check for void `subscriptions` in REALM_INIT.
Our `RealmInitAction` type is quite clear that `subscriptions` is
always present. That type is *meant* to represent our best knowledge
of what will come from the server, but it's important to verify.

Greg did some digging [1] and reported that the `|| initialState`
check appeared to be modeled on the solution to a problem that
didn't, after all, affect `subscriptionsReducer`:

"""
Looks like that originates in 462215c26. Following links, that was
from #2878 which fixed #2859. The latter issue was that, in a
_different_ reducer, we needed exactly this because the data in
question was about a recent feature that some servers didn't support
and so the property (namely `realm_user_groups`) would be missing.
"""

Ray also did some archaeological investigation to see if it was
reasonable to trust that the server has never handed out `null` or
another falsy value, and concluded [2],

"""
As of at least 7a930afa0 (2017-02-20), and probably much earlier, it
should always send at least `[]` for `subscriptions` so long as
`subscription` is specified as a desired event type in the
`/register` call. (I'm pretty sure that's true both at that commit
and in today's code; but there _are_ a lot of intermediate changes
that I haven't checked fully, so I'm less sure it's been true the
entire time.)
"""

[1]: https://github.com/zulip/zulip-mobile/pull/4046#discussion_r414888509
[2]: https://github.com/zulip/zulip-mobile/pull/4046#discussion_r414901766