async tests: Conform to Jest's API on actual vs. expected.
When a test fails, Jest's error messages lean quite hard on this
expectation.  Here's a failure from #4010:

  ● sleep (real) › waits for approximately the right number of milliseconds

    expect(received).toBeLessThanOrEqual(expected)

    Expected: <= 999
    Received:    1000

(Plus the "received" and "1000" are highlighted in red, and the
"expected" and "999" in green.)

The truth is that we expected a value >= 1000 and received 999,
so having it shown backward comes out pretty confusing.
Flip the arguments so Jest shows them the right way around.