webview css: Apply `font: -apple-system-body` to root element.
On Apple platforms, this causes the text to scale appropriately with
the system font-size settings.  On Android it has no effect, because
Chrome for Android just automatically handles this correctly.

It's puzzling that Apple seems to have arranged things so that these
accessibility settings don't work on web pages unless the developer
uses this arcane undocumented syntax to opt in.  But, so be it.

See tutorial here:
  https://dev.to/colingourlay/how-to-support-apple-s-dynamic-text-in-your-web-content-with-css-40c0

[greg: revised commit message]

Fixes: #4042