AuthScreen: Stop using doubled slashes in social auth login endpoints.
`.login_url` on objects in `external_authentication_method` has
always started with a leading slash, as far as I've seen. That means
we're constantly sending requests to the server that look wrong,
like

https://chat.zulip.org//accounts/login/social/google

. Production servers, apparently, are happy to pretend that it's a
single slash and handle it with no problems; my local server in dev
gives a 404. [1]

In any case, we shouldn't be using weird URLs like this at all; it's
likely that we wouldn't see bugs like this if we were using a proper
library for joining URLs, instead of string concatenation; that's
issue #4081.

[1]: https://github.com/zulip/zulip-mobile/issues/3567#issuecomment-622574435