typing status: Use all-users map at end of getCurrentTypingUsers, too.
Unlike the issue fixed in the last couple of commits, the situation
where this makes a difference seems unlikely to trigger unless
you're really trying -- the most likely case that comes to mind is
if you're PM'ing with a user, they're typing at you, then their
account is deactivated.

Still, fix it.  There's no fundamental reason why this code should
care about whether the users typing are deactivated, or a particular
flavor of bot, or what.

Because this data structure can contain cross-realm bots, this
change also causes the `UserOrBot` type to propagate through where
this data goes.