api: Create a binding to get temporary URL to access uploaded file.
The endpoint GET `/user_uploads/{realm_id_str}/{filename}` allows
for the generation of a temporary file URL that does not require
authentication. This endpoint was added in zulip/zulip@9e72fbe3f.

The goal basically is to get:

(a) a plain old URL that can be passed to a browser (not a URL
plus some cookies or other headers that need to be set).

(b) that gets the user the file immediately, without requiring a
fresh login.

(c) but doesn't include a secret with much wider powers (i.e.
their API key), which would risk a nasty surprise if they
innocently show someone else the file by sending them the link.

Commit-message-suggested-by: Greg Price <greg@zulipchat.com>