i18n types [nfc]: Make some object types explicitly inexact or exact.
The type bound on P is supposed to be inexact -- the idea is that
P probably has a bunch of other properties too (the other props of
the component) -- so make that explicit.

For the second argument to `$Diff`, I believe the actual behavior
of `$Diff` is unaffected by whether the argument is exact or inexact,
but it seems clearer to think about it as exact.  We also use exact
object types in the most complicated and carefully-thought-through
place where we use `$Diff`, which is our `react-redux.js` that
provides the type for `connect`.  So make those exact.