button text: Change "Log in with" to "Sign in with" everywhere.
Apple's Human Interface Guidelines doc on the "Sign in with Apple"
button [1] (to be added in an upcoming commit) is quite clear that,
for logging in, the text must say "Sign in with Apple", with that
exact capitalization and wording.

In order to keep the buttons consistent, change the others to match.

[1]: https://developer.apple.com/design/human-interface-guidelines/sign-in-with-apple/overview/buttons/.