expo-apple-auth libdef: Make identityToken, authorizationCode non-optional
The Apple docs are frustratingly vague, but it doesn't seem to be
the case that "no credential" and "yes credential, but no
`identityToken`/`authorizationCode`" are separate auth failure modes
that we have to distinguish.

As I theorize in an Expo issue, these fields being marked optional
does correspond to the Swift version of an Apple doc [1], where
they're also marked optional.

We still don't know exactly why that Apple doc has them marked
optional, but corresponding values in a different protocol, the one
you use on other platforms (i.e., not the native iOS flow), *are*
marked optional for a particular reason [2]. In particular, they
might be missing if you said you didn't want them. But...here, there
doesn't seem to be a way to say you don't want them. So, they must
be there.

Also, in the JavaScript layer of `expo-apple-authentication`, an
error is thrown if `identityToken` or `authorizationCode` is
missing. (See `signInAsync` in `build/AppleAuthentication.js`.)

[1]: https://github.com/expo/expo/issues/8446#issuecomment-634997122
[2]: https://github.com/expo/expo/issues/8446#issuecomment-635000576