deps: Remove `react-test-renderer`.
This would be useful to us if we were interested in doing snapshot
tests of our React components. Currently, we're not. If we add some
and bring this back, it should be in `devDependencies`.

Anecdotally, I've used this tool before, outside Zulip. I had
trouble with basic needs, for testing, like calling the component
instance's `setState` method. I'd recommend looking at Enzyme before
choosing to add this back in, but I don't know how either of these
tools might have changed since I last had a look.

From the doc [1]:

"""
Overview

This package provides a React renderer that can be used to render
React components to pure JavaScript objects, without depending on
the DOM or a native mobile environment.

Essentially, this package makes it easy to grab a snapshot of the
platform view hierarchy (similar to a DOM tree) rendered by a React
DOM or React Native component without using a browser or
[jsdom](https://github.com/tmpvar/jsdom).
"""

[1]: https://reactjs.org/docs/test-renderer.html