android: Kotlin-side code for receiving data from other apps.
This is part of the implementation of #117, to show up in the
"share" UI from other apps so the user can send a message with
the shared data.

This commit consists of the needed Android-native code, which uses
the relevant Android APIs to get the data and then the relevant
RN-on-Android APIs to send the data over to our main JS codebase.

On its own, this code doesn't yet do anything useful because we
don't have the JS code to listen for the data.  That's still in
development as #4124, and coming soon.

Because the feature doesn't yet work in this version, the manifest
elements to advertise it are commented out, so that it doesn't
actually get presented to users.

Also include the new dummy root React component, `SharingRoot`,
because the Kotlin-side code refers to it by name.  It exists in the
first place because 'Sharing' is linked with launching an activity
in the Android ecosystem.  But we don't always want to launch
`MainActivity` when receiving a share because it may cause two
instances of it to be opened simultaneously.  We can't check whether
the app is running before an activity launches.  So, we launch this
dummy component, then process the intent, identify whether the app
is running or not, and handle that as mentioned in the paragraph
above, and then quickly kill this dummy Activity.  All of this
happens fast enough that the component does not even get time to
render, so it's seamless.

[greg: split out Android-native portion as its own commit;
 adjusted commit message accordingly]