sharing: Add UI to receive and handle shares from other apps.
This completes most of the implementation of #117.  To enable the
feature as it currently exists, it's enough to uncomment the
commented-out bit of the manifest, so that Zulip appears in the
"sharing" UI when attempting to share something from another app.

The reason we don't yet enable/advertise the feature is that there
are some blocker bugs -- most notably, it crashes when you try to
use it.  That didn't use to happen, on a branch based on an older
version from master, but it does now after rebasing.  (The error
message suggests it's probably related to our recent upgrades of
the RN version we use.)

The 'react-navigation' code was suggested by Chris Bobbe.

Suggested-in-part-by: Chris Bobbe <cbobbe@zulip.com>