libdefs: Compose and use libdef for `react-native-webview` at 7.x.x.
The latest version FlowTyped has a libdef for is 6, unfortunately.

Ah, well. This is a best effort at replicating the types at
react-native-community/react-native-webview@c4001338c, tagged as
v7.6.0, the latest 7.x.

An improvement over the FlowTyped libdef is that WebView's props are
now exact. So, e.g., people will be warned not to use the
`useWebKit` prop, which was removed in 7.0.0.

We found it impractical to correctly express the fact that the
`WebView` component also accepts all the props that the `View`
component does. Certainly it seems impossible to do so by importing
things from React Native [1]. So, a partial reimplementation that
allows a `style` prop (typed as an inexact object).

See the entry in `docs/howto/libdefs.md` added in this commit for
more detail.

[1]: https://github.com/zulip/zulip-mobile/issues/3458#issuecomment-639859987