deps: Upgrade `react-native-webview` to 8.0.4.
The only declared breaking change [1] is the following:

"""
onNavigationStateChange now triggers with hash url changes
"""

We haven't been using that prop, so this is fine. (Even if we
were...I guess that could be a breaking change?)

Inspecting any changed TypeScript files in a local clone of
`react-native-webview`, between versions 7.6.0 and 8.0.4, suggests
that we don't have to make any changes to the libdef:

```
$ git diff v7.6.0..v8.0.4 -- *.ts
```

(No results.)

We'd like to go to the latest (10.2.3, currently), but
react-native-community/react-native-webview@bf1d64571, released in
v8.0.5 [2], bumps the `react` peer dependency to ^16.9.

We're using `react` at 16.8.6, since `react-native` 0.60.6 has
`react` as a peer dependency on 16.8.6 exactly; this gets bumped in
facebook/react-native@0ccedf396, released in v0.61.0.

So, effectively, this means `react-native-webview` v8.0.5 bumps the
React Native requirement to v0.61. Seems surprising that they didn't
mention that, but, oh well.

[1] https://github.com/react-native-community/react-native-webview#versioning

[2] https://github.com/react-native-community/react-native-webview/releases/tag/v8.0.5