deps: Use `@react-native-community/cameraroll`.
`CameraRoll` in React Native core is deprecated [1]. In the upcoming
RN v0.60 -> v0.61 upgrade, we see errors that it has in fact been
removed from `react-native`. We are to use
`@react-native-community/cameraroll` instead. So, do.

Very recently, the `saveToCameraRoll` method was removed in favor of
`save` [2]. The first argument is exactly the same; further config
is possible in later arguments if we want to.

Thankfully, this library has built-in Flow support, so we don't need
to make our own libdef. But the types haven't yet been updated to
reflect the removal of `saveToCameraRoll`; that should be coming
soon [3].

Also tell our Jest config to use Babel on
`node_modules/@react-native-community/cameraroll`, as it's not
compiled.

[1]: https://reactnative.dev/docs/0.60/cameraroll
[2]: https://github.com/react-native-community/react-native-cameraroll/pull/183
[3]: https://github.com/react-native-community/react-native-cameraroll/issues/184