ios: Remove ZulipMobileTests target and references.
This is a stub that's been hanging around for a long time without
being used. As long as we don't use it, it just causes confusion.

We made this decision [1] while considering some changes suggested
by the upgrade helper [2] during the RN v0.60 -> v0.61 upgrade. In
fact, Greg points out that what it's doing is in fact testing that
we render *the welcome screen of the hello-world app*. So it makes
sense to happen in the template app, as an example, but it's bound
to fail in our app.

The changes to the Xcode config were done entirely through the Xcode
UI: removing the target and various references to it (e.g., in "Edit
Scheme"), searching for the text "ZulipMobileTests" to see if we
were done yet.

[1]: https://github.com/zulip/zulip-mobile/pull/4150#pullrequestreview-429378027