libdefs: Copy new version of redux-mock-store libdef.
This version [1] is marked for a higher version of Flow than we're
using now (v0.104 and above), but there aren't any errors on our
Flow version (v0.98). So, take it.

This gets us a particular change to the Dispatch type [2] that goes
in the right direction...until we clobber it with our own Dispatch
that accounts for async redux-thunk actions, I suppose.

Ah well, there are a few minor changes that we might like to have,
and this is a later version, so, might as well.

[1]: https://github.com/flow-typed/flow-typed/blob/c4f47bdda/definitions/npm/redux-mock-store_v1.2.x/flow_v0.104.x-/redux-mock-store_v1.2.x.js
[2]: https://github.com/flow-typed/flow-typed/pull/3803

[cherry-picked from #4171]