libdefs: Hard-code redux-thunk support into redux-mock-store.
As foreshadowed in the preceding commits, we can't really expect a
FlowTyped-hosted libdef to have the flexibility to give us the right
types for Dispatch on the condition that we pass `thunk` as a piece
of middleware.

We're decidedly using `thunk`, so, hard-code support for it here by
allowing dispatch to be called with a redux-thunk async action.

We're pretty consistent with having correct types for such an
action, so not much is lost by typing its arguments loosely as
`Function, Function`, and from some attempts, it seems much easier
than nailing down something exactly right.

[cherry-picked from #4171]