mocks: Remove __mocks__/redux-mock-store.js.
It looks like this was added with the idea that we'd use it in more
places than we actually have. It adds a slight bit of convenience;
with this mock, we can avoid having to write
`configureStore([thunk])` a bunch of times.

But that's a very small amount of boilerplate.

The mock doesn't seem to have any other purpose that might be part
of a unit-testing strategy.

Removing it makes it possible to use the libdef without contorting
it more drastically than we already have.

[cherry-picked from #4171]