fetchActions tests: Fix buggy test.
This was clearly not testing what it said it was. The title is
correct; the "expect" was wrong, and some malformed input fulfilled
that wrong expectation. The live code is fine.

Turns out, we were using the actual string "HOME_NARROW" in
`state.fetching`, which led to `DEFAULT_FETCHING` to be used in
`getFetchingForNarrow`.

This actually wouldn't have been caught with type checking, since
the type for the keys in `state.fetching` is just a string, and
"HOME_NARROW" is a perfectly valid example, just as stringified JSON
is.

See discussion [1].

[1]: https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/fetchActions-test.2Ejs/near/912211