topic actions [nfc]: Rename fetchTopicsForStream.
The name 'fetchTopicsForActiveStream' is misleading because it
gives the impression that it internally determines the active
stream - but that is not so, and the stream must be supplied to it
as an argument. Also, it does not define what 'active' means.

Rename it to 'fetchTopicsForStream', that better describes its
interface.

Suggested-by: Chris Bobbe <cbobbe@zulip.com>