docs/release: Add reminder to sync from Transifex.
Historically I haven't been super consistent about doing this, which
is too bad because it means our users don't get the benefit of our
volunteer translators' work until later.  I do consistently look back
at this document when making a release, so this should help.

A good further improvement would be to reduce the number of steps
involved, with a bit of scripting -- in particular for uploading and
syncing across languages.  But as usual the first step in automation
is to just write the steps of the task down at all.