narrowsSelectors tests [nfc]: Remove confusing `expectedState` variables.
`expectedState` makes sense in tests for reducers, but not really
for selectors. The *input* of a selector is the GlobalState, but the
output is...whatever we want to do with that input. Sometimes we
just pick things directly from the state; sometimes we do that and
then nudge it around until it's in a shape we want.

Also rename `anchor` to the more general `result`; I'm not really
sure what "anchor" is meant to refer to in each of these instances.