android [nfc]: Tighten return type on a converted-from-Java method.
This was a Java method returning String, and so the conversion to
Kotlin made it return `String?` because the Java type String
includes null.

But in fact it never does return null, and it'd be contrary to the
intent of the method if it ever did.  So we can improve the type by
ruling that out.