android: Skip sharing intent when it's reused after quitting.
The android recents drawer is basically a log of intents. When we
finish sending a 'share' from our app, the original intent that
launched the sharing UI still appears in this log. So, detect if the
app was launched from the recents drawer after quitting, and if so,
don't show the sharing UI.

This prevents the app from showing the sharing UI when it's resumed
after completing/cancelling a share.