exampleData [nfc]: Find a simpler Flow inexact-spread workaround.
Ideally this code would look like:

    const { stream: streamInner = stream, ...extra } = args ?? {};

But that produces a Flow warning where `extra` is used:

  Cannot determine the type of [object literal] [1] because it
  contains a spread of inexact [rest of object pattern] [2]. Being
  inexact, [rest of object pattern] [2] might be missing the types of
  some properties that are being copied. Perhaps you could make it
  exact?  (inexact-spread)

Which is a fine point in general, but puzzling here, because the
object type in `args` is exact, not inexact.

The pattern in the existing version of this code avoids the warning,
but that seems to come from the use of the conditional making Flow
lose track of what's happening -- if I try using `args ?? {}` in a
single unconditional assignment instead, I get the same warning.

On further experimentation, the issue is the case of an undefined
`args`, and that the object literal `{}` which we use in that case
is being given inexact type `{}` instead of exact type `{||}`.
Using a nonempty object literal instead fixes the problem.
Happily, there is a natural such object to use that leads to
the same behavior.