tools: Add a run-android script.
As the comment explains, `react-native run-android` is unable to run
our app in debug mode with the correct app ID and activity name.

We've had a series of hacks that, first, convince it of the right app
ID (which it misdetects on its own), and then provide a dummy activity
with the name it expects: com.zulipmobile.debug.MainActivity.  See
commit 0d16b604b, which introduced these hacks and explains the issues
in `react-native run-android` that led to them.

But having a dummy activity leads to surprising, subtly broken,
behavior when combined with launching the app through other means that
use the normal activity name instead of the dummy one: the launcher,
or the new "sharing" feature.  See chat discussions involving some
very puzzled debugging:
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/need.20for.20conditional.20in.20navReducer's.20.60rehydrate.60.20for.20Sharing/near/946142
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/double.20initialization/near/947471

So we're going to take that dummy activity out.

This script provides an alternative to `react-native run-android`, one
that works correctly with our actual app ID and main-activity name.
In subsequent commits, we'll change references to that command to
point to this script instead, and remove the dummy activity and the
other hacks.