android: Cut hacks that catered to `react-native run-android` brokenness.
Even with these hacks, the behavior was still subtly broken in certain
cases because the activity name was different from the "real" main
activity name, the one used by the launcher and by the upcoming
sharing feature.  Time to give up on adapting to the broken tool, and
instead stick to the reasonable thing in the app (no duplicate dummy
activities) and adapt our tools to that.

In the preceding commits, we've added the script `tools/run-android`
to replace `react-native run-android` and switched our recommendations
to point to the new script instead of the latter.  Now drop the hacks,
in particular the dummy activity, so that it's no longer possible to
encounter the broken behavior they enabled.