Update the Zulip logo, and add a script to do it.
Hopefully the script will make it easier to regenerate everything if
there's a v2.1 or v3 of the logo in the future, or if we want to tweak
the graphical variation of the logo we use in a particular context, or
just to add more sizes as Apple comes out with new hardware and
requires particular new odd sizes for the app icon there.

A few changes, other than simply using the new v2 logo:

 * On Android we include a higher-res version of the launcher icon
   than we had before: xxxhdpi.  Nominally this is for a display
   density of 640ppi, which is beyond any real device; but there are
   devices with more than the 480ppi that corresponds to the next
   density down, and they'll generally prefer to scale the higher-res
   image down than vice versa.

   * For the debug variant, we also copy the xxhdpi version over to
     xxxhdpi.  Otherwise, on a debug build, the launcher on my
     emulated Pixel 2 ends up preferring the higher-res version rather
     than the one for the right variant.

 * For the app icon on iOS, we drop a bunch of sizes that we had lying
   around but that aren't recommended in the current documentation
   from Apple.  In particular this includes a bunch of "1x" images for
   low-density screens... which turns out to mean the iPhone 3GS and
   older, which haven't been supported since iOS 7.

   * We also normalize the filenames, which has some variation in them
     that seems to be meaningless.

There'll be one additional step to deploy this, beyond the usual
builds and uploads: the icon in the Play Store needs to be updated
through the Play Console.  For the App Store this will IIUC be taken
care of by the usual upload process, because the App Store icon is
part of the "icon set" and gets built into the artifact we upload;
but for the Play Store it isn't.

Fixes: #4200