deps: Slight upgrade of all Jest-related dependencies.
Done with the following command:

yarn upgrade -- $(tools/deps | grep jest) && yarn yarn-deduplicate

(See discussion [1].)

Kicking the tires, there are just a few issues that pop up, so we
fix those in the preceding commits.

I noticed one possibly interesting change: when mikehardy/jetifier
runs on postinstall, it always tells us how many files to
"forward-jetify" it finds in node_modules (and presumably applies
its string substitutions on all of them; see e433197cf). Before this
commit, it was 2178, and now it's 1952: a reduction by over 200!
This is a good sign. The third-party libraries we use are still
evolving.

This observation would be particularly interesting if I were going
to further investigate the error in @rnc/netinfo about the Android
Support Library, which prompted me to upgrade that library in a
recent commit. Thankfully, upgrading it was painless and made the
error go away, and we should be on the latest version we can anyway.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/changes.20in.20generatedEs3.2Ejs/near/935879