Revert "translation context: Eliminate `key` hack (the last one!)."
This reverts commit 68ca7468.

We were too quick with this, unfortunately. The `FormattedMessage`
component, straight from `react-intl`, is a consumer of
`IntlProvider`, using the old Context API. So we still need the
`key` hack.

See discussion [1].

We should then move forward with upgrading `react-intl` to v3, which
uses the new Context API. With that, we won't need the `key` hack.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Language.20picker.20regression/near/974627