Revert "deps: Upgrade ESLint to ^6.8.0, and related deps."
This reverts commit 01593b3af.

We should redo the upgrade when we can take a version of
`prettier-eslint` in which `core-js` is re-introduced as a
dependency; prettier/prettier-eslint#348 is the closest issue for
that. The commands listed in 01593b3af should make it easy to redo
the upgrade mechanically, and without having to think about rebase
conflicts.

The version (11.0.0) of `prettier-eslint` that we got in 01593b3af
was affected by a bug with symptoms quite similar to ones we see
with the Prettier VSCode extension (a.k.a. `esbenp.prettier-vscode`)
at version 5 and above [1]. We'll have to resolve the problem with
the VSCode extension at some point (see discussion for a likely way
we'll do that), since we can't assume it will always be fine to stay
below version 5, and version 5 causes problems because of an
intentional design change.

But, from experimentation, it appears that this time the problem is
just caused by `prettier-eslint` not having the right version of
`core-js`. At 10.1.0, the latest version before the problem starts
happening, we can remove
`node_modules/prettier-eslint/node_modules/core-js`, and we see the
problem. Version 10.1.1, which ships with the problem, removed
`core-js` as a dependency, and an issue with a different bad symptom
(prettier/prettier-eslint#348) was soon filed, and a PR
(prettier/prettier-eslint#359) came along to add `core-js` back in.

We can't just back up to version 10.1.0, because then we're at risk
of it not working with ESLint 6 [2]. So, just revert the entire
ESLint 6 upgrade.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/SOLVED.3A.20Prettier.20auto-formatting/near/893164
[2] https://github.com/prettier/prettier-eslint/pull/338#issuecomment-632908721