typing status [nfc]: Rename typingStatusExpiryLoop and fold into module.
The existing name was very confusing because it sounded like something
that would clear the typing indicators/notifications *now*, not like
something that would start a loop to repeatedly check back in the
future to expire them as appropriate.  Confusing not only in potential
but in actual confusion; see:
  https://github.com/zulip/zulip-mobile/pull/4216#discussion_r468105975
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Typing-state.20client.20timer/near/986619

This also makes a good time to fold it into a module with a more
general name, so that the filename isn't tied to the function's name.