render test: Use representative `realm` value for `messageTypingAsHtml`.
It looks like this was trying to test that the sequence `&<` gets
escaped correctly. But we're already testing that for the user's
`email` and `avatar_url`, and I don't think we expect a realm URL to
contain that sequence.

When we start making `getFullUrl` use the URL constructor in an
upcoming commit, this test would have thrown an invalid URL error.
At that commit, we'll be sure to test manually, noting any invalid
URL errors we see at runtime, and address them.