redux: Add jitsi server url to realm state.
To prepare for rendering a button that inserts video call links in the
ComposeBox/ComposeMenu, we need to store the jitsi server url in the
realm slice of the redux state and add a selector for accessing it.
Since not all zulip clients will be interfacing with a zulip server that
supports this feature, the field can either be of type string or null.