flow: Follow a React Native change.
Part of the RN v0.61 -> v0.62 changes to the template app [1],
corresponding to facebook/react-native@46090b0e5. I don't find that
this must happen at or after the main upgrade commit; it's a bugfix
that makes a pattern less restrictive than it was before, and it's
not obviously attached to any changes in `react-native` in the
upgrade.

[1] https://react-native-community.github.io/upgrade-helper/?from=0.61.5&to=0.62.2